fix: resolve 23 PR review comments (4 critical, 16 major, 3 minor)
Critical fixes:
- Fix SIMD guards: change Avx.IsSupported to Avx2.IsSupported for all backward
  kernels using ExpApprox256 (Swish, Mish, Softplus, SELU) — ExpApprox256 uses
  Avx2.ConvertToVector256Int32 which crashes on AVX-only CPUs
- Fix SplitBackward: use savedState axis/start to scatter chunk gradient back to
  correct position in original input shape (was incorrectly accumulating chunk-shaped
  gradient directly)
- Fix Concatenate recording: clone tensors array before storing on tape to prevent
  array element replacement
- Fix BatchMatMul double recording: 3D case delegates to BatchMatMul which already
  records — remove duplicate RecordBinary that doubled gradients

Major fixes:
- AutogradFunction: clone inputs array before storing on tape
- ConvTranspose2D: clone stride/padding arrays before saving to tape
- Floor/Ceiling: use zero gradient (SignBackward) instead of STE — floor/ceil are
  piecewise constant, not quantization ops
- CompiledBackwardGraph: snapshot tape entries list + add ValidateInputVersions
- CompileBackward: reject non-scalar losses with clear error message
- GradientAndUpdate: throw on gradient-size mismatch instead of partial update
- LeakyReLU GPU: save alpha as double (not float) to match backward contract
- TensorPool.Clear: document thread-local scope behavior
- Benchmarks: skip in CI with [Fact(Skip=...)] — run manually only
- Fix compiled backward tests to reduce loss to scalar before CompileBackward

1465 tests pass, 0 failures, 61 skipped (GPU + benchmarks)

+++ src/AiDotNet.Tensors/Engines/DirectGpuTensorEngine.cs
@@ -10876,7 +10876,7 @@ public partial class DirectGpuTensorEngine : CpuEngine, ITensorLevelEngine, IDis
             var gpuOutput = new Tensor<T>(result, tensor.Shape._dims);
             Autodiff.DifferentiableOps.RecordUnary("LeakyReLU", gpuOutput, tensor,
                 Autodiff.BackwardFunctions<T>.LeakyReLUBackward,
-                savedState: new object[] { alphaFloat });
+                savedState: new object[] { (double)alphaFloat });
             return gpuOutput;
         }
         catch (Exception)
